feat (school detail) : Upgrade School Details Page from Nuxt 2 to Nuxt 3
