+++ nuxt.config.js
@@ -176,7 +176,7 @@ export default defineNuxtConfig({
       proxy: "https://core.gamatrain.com/api/v1/**",
     },
     "/api/v2/**": {
-      proxy: "https://api.gamaedtech.com/api/v1/**",
+      proxy: "https://sandbox.gamaedtech.com/api/v1/**",
     },
     "/uploads/**": {
       proxy: "https://core.gamatrain.com/uploads/**",
